# services/backend/smart-health-platform/packages/api-contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Health & Supply Chain Resilience — API Contracts
  version: "1.0.0"
  description: >
    Authoritative single source of truth for the PHC-facing REST surface,
    offline-first delta sync engine, and authentication contracts.
    Consumed by PHC Portal, Governance Portal, and Core Backend.
servers:
  - url: http://localhost:3000
    description: Local development backend
  - url: https://api.healthresilience.gov/v1
    description: Production National DC Gateway
security:
  - bearerAuth: []
paths:
...
  /api/v1/phc/{phcId}/facility:
    get:
      summary: Get facility profile and current operational capacity
      operationId: getFacility
      tags: [Facility]
      parameters:
        - $ref: '#/components/parameters/phcId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Facility'
    put:
      summary: Update facility operational parameters (beds, oxygen)
      operationId: updateFacility
      tags: [Facility]
      parameters:
        - $ref: '#/components/parameters/phcId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityUpdateInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Facility'
  /api/v1/phc/{phcId}/inventory:
    get:
      summary: List inventory batches with FEFO sorting and expiry tracking
      operationId: listInventory
      tags: [Inventory]
      parameters:
        - $ref: '#/components/parameters/phcId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryBatch'
    post:
      summary: Receive / add an inventory batch
      operationId: addInventoryBatch
      tags: [Inventory]
      parameters:
        - $ref: '#/components/parameters/phcId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryBatchInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryBatch'
...
  /api/v1/phc/{phcId}/equipment:
    get:
      summary: List critical equipment status (ventilators, oxygen concentrators, cold chain)
      operationId: listEquipment
      tags: [Equipment]
      parameters:
        - $ref: '#/components/parameters/phcId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EquipmentItem'
...
  /api/v1/phc/{phcId}/footfall:
    get:
      summary: Retrieve patient footfall counts by category
      operationId: getFootfall
      tags: [Footfall]
      parameters:
        - $ref: '#/components/parameters/phcId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FootfallRecord'
    post:
      summary: Append patient footfall observation (append-only)
      operationId: recordFootfall
      tags: [Footfall]
      parameters:
        - $ref: '#/components/parameters/phcId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FootfallInput'
      responses:
        '201':
          description: Footfall recorded
  /api/v1/phc/{phcId}/requests:
    get:
      summary: List resource replenishment and emergency requests
      operationId: listRequests
      tags: [Requests]
      parameters:
        - $ref: '#/components/parameters/phcId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResourceRequest'
    post:
      summary: Create replenishment or operational resource request
      operationId: createRequest
      tags: [Requests]
      parameters:
        - $ref: '#/components/parameters/phcId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceRequestInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceRequest'
  /api/v1/phc/{phcId}/emergency:
    post:
      summary: Fast-path emergency incident report (CRITICAL severity)
      operationId: reportEmergency
      tags: [Emergency]
      parameters:
        - $ref: '#/components/parameters/phcId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmergencyReportInput'
      responses:
        '201':
          description: Emergency alert generated and broadcast immediately
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertItem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    phcId:
      name: phcId
      in: path
      required: true
      description: Target PHC facility UUID
      schema:
        type: string
        format: uuid
  schemas:
...
    Facility:
      type: object
      required: [id, name, district_id, state_id, total_beds, occupied_beds, oxygen_cylinders]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        district_id: { type: string, format: uuid }
        state_id: { type: string, format: uuid }
        total_beds: { type: integer }
        occupied_beds: { type: integer }
        oxygen_cylinders: { type: integer }
        status: { type: string }
    FacilityUpdateInput:
      type: object
      properties:
        total_beds: { type: integer }
        occupied_beds: { type: integer }
        oxygen_cylinders: { type: integer }
    InventoryBatch:
      type: object
      required: [id, phc_id, medicine_id, batch_no, remaining_qty, expiry_date]
      properties:
        id: { type: string, format: uuid }
        phc_id: { type: string, format: uuid }
        medicine_id: { type: string, format: uuid }
        medicine_name: { type: string }
        batch_no: { type: string }
        remaining_qty: { type: integer }
        minimum_threshold: { type: integer }
        expiry_date: { type: string, format: date }
    CreateInventoryBatchInput:
      type: object
      required: [medicine_id, batch_no, remaining_qty, expiry_date]
      properties:
        medicine_id: { type: string, format: uuid }
        batch_no: { type: string }
        remaining_qty: { type: integer, minimum: 1 }
        minimum_threshold: { type: integer, default: 10 }
        expiry_date: { type: string, format: date }
...
    EquipmentItem:
      type: object
      required: [id, phc_id, equipment_type, status]
      properties:
        id: { type: string, format: uuid }
        phc_id: { type: string, format: uuid }
        equipment_type: { type: string }
        serial_no: { type: string }
        status: { type: string }
        last_inspected_at: { type: string, format: date-time }
...
    FootfallRecord:
      type: object
      required: [time, phc_id, category, count]
      properties:
        time: { type: string, format: date-time }
        phc_id: { type: string, format: uuid }
        category: { type: string }
        count: { type: integer }
    FootfallInput:
      type: object
      required: [category, count]
      properties:
        category: { type: string, enum: [general_opd, maternal_child, emergency, chronic_followup, infectious] }
        count: { type: integer, minimum: 1 }
        time: { type: string, format: date-time }
    ResourceRequest:
      type: object
      required: [id, phc_id, request_type, priority, status, created_at]
      properties:
        id: { type: string, format: uuid }
        phc_id: { type: string, format: uuid }
        request_type: { type: string }
        priority: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        decided_at: { type: string, format: date-time }
    ResourceRequestInput:
      type: object
      required: [request_type, priority, items]
      properties:
        request_type: { type: string, enum: [replenishment, emergency, equipment_repair] }
        priority: { type: string, enum: [low, normal, high, critical] }
        items:
          type: array
          items:
            type: object
            required: [medicine_id, requested_qty]
            properties:
              medicine_id: { type: string, format: uuid }
              requested_qty: { type: integer, minimum: 1 }
    EmergencyReportInput:
      type: object
      required: [alert_type, title, description]
      properties:
        alert_type: { type: string, enum: [mass_casualty, disease_outbreak, oxygen_failure, extreme_weather] }
        title: { type: string }
        description: { type: string }
    AlertItem:
      type: object
      required: [id, alert_type, severity, status, created_at]
      properties:
        id: { type: string, format: uuid }
        phc_id: { type: string, format: uuid }
        district_id: { type: string, format: uuid }
        state_id: { type: string, format: uuid }
        alert_type: { type: string }
        severity: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
...